Rename Registration service as Enterprise
As there is no product registration anymore, this service exists solely
to communicate with the on-prem Enterprise Manager. Consequently, it is
far more illuminating to users to call this service [enterprise].
